requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI & ML
openai==1.3.0
langchain==0.0.350
pypdf2==3.0.1
python-docx==1.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Mapping & Geolocation
geopy==2.4.0
googlemaps==4.10.0

# PDF Generation
reportlab==4.0.7
fpdf2==2.7.6

# Azure
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
pydantic-settings==2.1.0


